Предварительно сделал postDataToUser. Тестирую

+++ src/app/utils/functions/getAmountByType.ts
@@ -1,8 +1,10 @@
 const getAmountByType = (amount: string, cardType: string) => {
   if (cardType === "expense") {
-    return -parseInt(amount, 10);
+    //@ts-ignore
+    return -parseFloat(amount, 10);
   }
-  return parseInt(amount, 10);
+  //@ts-ignore
+  return parseFloat(amount, 10);
 };
 
 export default getAmountByType;
